Use ssize_t instead of int in src/syntax/highlight.c

+++ src/syntax/highlight.h
@@ -7,7 +7,7 @@
 #include "../terminal/color.h"
 
 HlColor **hl_line(Buffer *b, const LineRef *lr, size_t line_nr, bool *next_changed);
-void hl_fill_start_states(Buffer *b, int line_nr);
+void hl_fill_start_states(Buffer *b, size_t line_nr);
 void hl_insert(Buffer *b, size_t first, size_t lines);
 void hl_delete(Buffer *b, size_t first, size_t lines);
